+++ src/neo-express/Commands/CheckPointCommand.Create.cs
@@ -20,9 +20,6 @@ namespace NeoExpress.Commands
             [Option]
             private bool Force { get; }
 
-            [Option]
-            private bool Online { get; }
-
             private async Task<int> OnExecuteAsync(CommandLineApplication app, IConsole console)
             {
                 try
@@ -50,22 +47,9 @@ namespace NeoExpress.Commands
                         throw new Exception("Checkpoint create is only supported on single node express instances");
                     }
 
-                    if (Online)
-                    {
-                        var uri = chain.GetUri();
-                        var result = await NeoRpcClient.ExpressCreateCheckpoint(uri, filename)
-                            .ConfigureAwait(false);
-                        console.WriteResult(result);
-                    }
-                    else
-                    {
-                        var blockchainPath = chain.ConsensusNodes[0].GetBlockchainPath();
-
-                        BlockchainOperations.CreateCheckpoint(
-                            chain, blockchainPath, filename);
-                    }
-
-                    console.WriteLine($"created checkpoint {Path.GetFileName(filename)}");
+                    var result = await BlockchainOperations.CreateCheckpoint(chain, filename)
+                        .ConfigureAwait(false);
+                    console.WriteLine(result);
 
                     return 0;
                 }